Added special space key handling in TextArea
The goal is to replicate the behavior of the standard TextBox control. Indeed TextBox handles the key space in a special way because space is also a 'control' key which can be used for validation for example.
For example without this special handling, the TextEditor does not work correctly it one of its ancestor handles the space KeyDown event while a TextBox would work just fine.
This implementation is only an attempt at mimicking the TextBox behavior, but since there is no documentation on the subject it is may not be fully correct.